alternative tables for FPU
the solutions provided by the following tables are entirely optional, and not required for the functionality of emu68 IMHO. it can be used as reference and or actually implemented, I hope these give a better understanding of the second word of the FPU instruction. rev3 is not done yet but I will be at that one in the near future.